% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{bar_race_animation}
\alias{bar_race_animation}
\title{Returns an animated bar graph}
\usage{
bar_race_animation(data)
}
\arguments{
\item{data}{tibble with columns `Value` (the y value for each bar), `name` (the label for each bar), `label` (the colour label for each bar), `colour` the colour for each bar, `Year` (time variable)}
}
\description{
Returns an animated bar graph
}
