% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{built_up_graph_for}
\alias{built_up_graph_for}
\title{Graph built-up area}
\usage{
built_up_graph_for(city_name, s, buffer = 30, ds = 5)
}
\arguments{
\item{city_name}{name of city or sf object with columns `name` and custom location}

\item{s}{stars object with buit data}

\item{buffer}{radius around central city location in km}

\item{ds}{downsample factor when plotting data}
}
\value{
a ggplot2 object
}
\description{
Graph built-up area
}
