% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{density_plot_series}
\alias{density_plot_series}
\title{Plot city density series}
\usage{
density_plot_series(
  location,
  years = c("1975", "1990", "2000", "2015"),
  radius_km = 30,
  max_density = NULL
)
}
\arguments{
\item{location}{a location class `sf` point geometries and `name` field}

\item{years}{list of years to include in the plot}

\item{radius_km}{radius of plots}

\item{max_density}{}
}
\value{
ggplot2 object
}
\description{
Plot city density series
}
