% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_city_locations}
\alias{get_city_locations}
\title{Get location data for city from maps package}
\usage{
get_city_locations()
}
\value{
a tibble with matching city location
}
\description{
Get location data for city from maps package
}
