% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_GHS2023A_for}
\alias{get_GHS2023A_for}
\title{Get GHS 2023A population and built data}
\usage{
get_GHS2023A_for(
  geo = NULL,
  type = c("BUILT_S", "BUILT_V", "BUILT-V_NRES", "POP", "BUILT_H_AGBH", "BUILT_H_ANBH"),
  resolution = c("100", "1000"),
  years = c("1975", "1980", "1985", "1990", "1995", "2000", "2005", "2010", "2015",
    "2020"),
  timeout = 10000,
  version = "V1-0",
  layer = NULL,
  base_path = getOption("custom_data_path")
)
}
\arguments{
\item{geo}{cut down to geography (will include data in buffer around geography)}

\item{type}{type of the sercies}

\item{resolution}{resolution}

\item{years}{year for the data series}

\item{timeout}{timeout for data download}

\item{version}{version of the dataset}

\item{layer}{optional layer}

\item{base_path}{local path to cache GHSL data}
}
\value{
stars object for ppulation data
}
\description{
Get GHS 2023A population and built data
}
