% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_GHS_30_built_data}
\alias{get_GHS_30_built_data}
\title{Get a reference to GHS built data, download if needed}
\usage{
get_GHS_30_built_data(
  base_path = getOption("custom_data_path"),
  timeout = 10000
)
}
\arguments{
\item{base_path}{Local path to store data}

\item{timeout}{timeout for data download}
}
\value{
stars object for built data
}
\description{
Get a reference to GHS built data, download if needed
}
