% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_GHS_built_data}
\alias{get_GHS_built_data}
\title{Get GHS land use data}
\usage{
get_GHS_built_data(geo = NULL, resolution = c("30", "250", "1K"),
  year = c("1975", "1990", "2000", "2014"))
}
\arguments{
\item{geo}{cut down to geography (will include data in buffer around geography)}

\item{resolution}{resolution}

\item{year}{year for the data series}
}
\description{
Get GHS land use data
}
