% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_GHS_for}
\alias{get_GHS_for}
\title{Get GHS populatin data}
\usage{
get_GHS_for(geo = NULL, resolution = c("250", "1k"), year = c("1975",
  "1990", "2000", "2015"))
}
\arguments{
\item{geo}{cut down to geography (will include data in buffer around geography)}

\item{resolution}{resolution}

\item{year}{year for the data series}
}
\description{
Get GHS populatin data
}
