% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{get_population_weighted_data}
\alias{get_population_weighted_data}
\title{Get population weighted data for cities and years}
\usage{
get_population_weighted_data(cities, years = "2015",
  max_radius_km = 30, resolution = "250")
}
\arguments{
\item{cities}{list of cities of class `sf` point geometries and `name` field}

\item{years}{list of years, valid values are `1975`, `1990`, `2000`, `2015`}

\item{max_radius_km}{radius to compute the densities}

\item{resolution}{1k or 250 metre grid}
}
\description{
Get population weighted data for cities and years
}
