% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\docType{data}
\name{ghs_attribution}
\alias{ghs_attribution}
\title{Attribution for data}
\format{An object of class \code{character} of length 1.}
\usage{
ghs_attribution
}
\description{
Attribution for data
}
\keyword{datasets}
