% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_densities.R
\name{pop_weighted_density_for}
\alias{pop_weighted_density_for}
\title{populatin weighted density}
\usage{
pop_weighted_density_for(location, max_radius_km = 40, year = "2015",
  type = "250")
}
\description{
populatin weighted density
}
